requirements.txt
# Stock Screener Dependencies
# Python 3.8+

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0

# Stock market data fetching
yfinance>=0.2.28

# YAML configuration file support
PyYAML>=6.0

# Backtesting Framework Dependencies
pandas_ta>=0.3.14b    # Technical indicators
pyarrow>=12.0.0       # Parquet support
tqdm>=4.66.0          # Progress bars
